Added feature of getting the new current episode
