+++ http/sourcegrabber.h
@@ -3,40 +3,17 @@
 
 #include <string>
 #include <iostream>
+#include <fstream>
+#include <vector>
+
 #include <QObject>
 #include <QtNetwork/QNetworkAccessManager>
 #include <QtNetwork/QNetworkReply>
 #include <QEventLoop>
 #include <QString>
 #include <QDebug>
-#include <fstream>
-#include <vector>
 
-struct Date
-{
-    int day;
-    int month;
-    int year;
-};
-
-struct TvSeries {
-    int id;
-    std::string name;
-    std::string image;
-    std::string link;
-};
-
-struct Episode
-{
-    int number;
-    Date airDate;
-};
-
-struct Season
-{
-    int number;
-    std::vector<Episode> episodes;
-};
+#include <core/tvtrackertypes.h>
 
 class SourceGrabber {
 public:
